Remove false note about limitation
This statement in the README is not true. I was thinking of _building_ messages,
where the nesting is indeed a bit awkward. How should we shorten the following
code?

nl80211_schema.build({"NL80211_ATTR_KEY": {"NL80211_ATTR_KEY_DEFAULT": 1}})

We can write this:

nl80211_schema.build(key={"NL80211_ATTR_KEY_DEFAULT": 1})

But it's not pretty to mix and match the Python and full-length names like
that..